refactor: move port change test from testbench
