+++ yarn-project/p2p/testbench/configurations/normal-degree-20-nodes.json
@@ -1,0 +1,14 @@
+{
+  "numberOfClients": 20,
+  "maxPeerCount": 30,
+  "gossipsubFloodPublish": false,
+  "debugDisableColocationPenalty": true,
+  "gossipsubInterval": 700,
+  "gossipsubD": 6,
+  "gossipsubDlo": 4,
+  "gossipsubDhi": 12,
+  "gossipsubDLazy": 6,
+  "peerCheckIntervalMS": 5000,
+  "gossipsubMcacheGossip": 12,
+  "gossipsubMcacheLength": 12
+}
